Add pre-generation of manuals archive for download
The dynamic generation and streaming of the archive was not reliable
so we added the ability for a maintainer to kick off the generation
of a staticly served ZIP file. This required more storage, so we also
had to add a data volume. All data was moved off to this larger AWS
volume significantly freeing up the root filesystem. Nginx had to be
modified to serve the static ZIP file, and the docker-compose file
had to be adjusted to point to the new volume locations.