Archive and Notification Changes
Modify the archive generation so that a date is attached to the archive
file to help users tell if they have the latest version or not. Make
changes to the server and the front end to display the archive date
information.

Additionally, add a Slack messaging component for when new files are
uploaded as well as when new archives are generated. These messages go
to a dedicated #manuals channel.